+++ demo/tests/test_core.py
@@ -161,48 +161,51 @@ class AutocompleteModelTest(TestCase):
     def test_final_queryset_person(self):
         instance = AutocompletePerson()
         # No items() called, None
-        self.assertTrue(instance._final_queryset is None)
+        self.assertTrue(instance.final_queryset is None)
 
         # Called items(), empty queryset
         items = instance.items()
-        self.assertFalse(instance._final_queryset is None)
-        self.assertEqual(instance._final_queryset.count(), 0)
+        self.assertFalse(instance.final_queryset is None)
+        self.assertEqual(instance.final_queryset.count(), 0)
 
         # Regular search, normal queryset
         items = instance.items(query="ali")
         self.assertEqual(len(items), 4)
-        self.assertTrue(instance._final_queryset)
-        self.assertEqual(instance._final_queryset.count(), 4)
+        self.assertTrue(instance.final_queryset)
+        self.assertEqual(instance.final_queryset.count(), 4)
 
         # Paginated queryset
         instance = AutocompletePerson(page_size=2)
         items = instance.items(query="ali")
         self.assertEqual(len(items), 2)
-        self.assertTrue(instance._final_queryset)
-        self.assertEqual(instance._final_queryset.count(), 2)
+        self.assertTrue(instance.final_queryset)
+        self.assertEqual(instance.final_queryset.count(), 2)
+        self.assertEqual(instance.final_raw_queryset.count(), 4)
 
     def test_final_queryset_personqueryset(self):
         instance = AutocompletePersonQueryset()
         # No items() called, None
-        self.assertTrue(instance._final_queryset is None)
+        self.assertTrue(instance.final_queryset is None)
 
         # Called items(), empty queryset
         items = instance.items()
-        self.assertFalse(instance._final_queryset is None)
-        self.assertEqual(instance._final_queryset.count(), 0)
+        self.assertFalse(instance.final_queryset is None)
+        self.assertEqual(instance.final_queryset.count(), 0)
 
         # Regular search, normal queryset
         items = instance.items(query="alice")
         self.assertEqual(len(items), 4)
-        self.assertTrue(instance._final_queryset)
-        self.assertEqual(instance._final_queryset.count(), 4)
+        self.assertTrue(instance.final_queryset)
+        self.assertEqual(instance.final_queryset.count(), 4)
 
         # Paginated queryset
         instance = AutocompletePersonQueryset(page_size=2)
         items = instance.items(query="alice")
         self.assertEqual(len(items), 2)
-        self.assertTrue(instance._final_queryset)
-        self.assertEqual(instance._final_queryset.count(), 2)
+        self.assertTrue(instance.final_queryset)
+        self.assertEqual(instance.final_queryset.count(), 2)
+        # Non-paginated queryset, full results
+        self.assertEqual(instance.final_raw_queryset.count(), 4)
 
 
 class AutocompletePersonTest(TestCase):